redefine base-install.d hooks
live-installer needs these hooks to run after the live system is copied in.
base-installer can get by with running them after deboostrap, and the
existing hooks will work ok if the system is already there.

So, redefine the base-install.d hooks to run after the base system is
extacted, but before the installed system is used in any way (such as
using apt to install additional stuff).

r53756