Start udev before running init
In order to solve the #484366, udev needs to be started before we start
busybox init as we don't want to create every /dev/ttyXXXX devices
manually.

udev-udeb (>= 1.125-6) moved the startup script from
/lib/debian-installer-startup.d/S02udev to
/lib/debian-installer/start-udev.  The later is called directly in
the initial init script.

Starting udev requires /sys to be available, and we now mount it
as we do for /proc.

udev already needed be available in order to make the installer boot, so
rootskel now depends on udev-udeb.  We can thus avoid testing for the
presence of udevd on the initrd.

As rootskel depends on udev-udeb, there is no further need to
explicitely udev-udeb to pkg-lists/base.

r55293