Exclude X.Org modules from library reduction
This is not an ideal solution, but workable for now. Exclusion is
needed because mklibs is not able to find symbols used in X.Org
modules that are provided by the /usr/bin/Xorg binary.

The risk with this solution is that symbols from other libraries
(e.g. libc6) used in an X.Org module but not used elsewhere could
now be reduced away even though they are actually needed.

r62790