Add Hyper-V modules to netboot and cdrom
Commit cd006086fa5d91414d8ff9ff2b78fbb593878e3c:

ata_piix: defer disks to the Hyper-V drivers by default

in mainline kernel and backported in debian kernel, it's not possible to use
ata_piix to access to hyper-v disk storage, which mean it's not possible
to install debian.
This commit is disabling ata_piix because the disk can be accessed with the
hyperv storage module. So, to install debian, the hyperv udeb is needed.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>
Signed-off-by: Cyril Brulebois <kibi@debian.org>